Fixed problem with event listener types and more
