+++ src/contract.ts
@@ -6,7 +6,6 @@ export type Map<T> = myra.Map<T>
  * Component types
  */
 export interface ComponentSpec<TState, TProps> extends myra.ComponentSpec<TState, TProps> { }
-export interface ComponentContext<TState, TProps> extends myra.ComponentContext<TState, TProps> { }
 export type ComponentFactory<T> = myra.ComponentFactory<T>
 
 export type OnMount<TState, TProps> = myra.OnMount<TState, TProps>
@@ -22,7 +21,9 @@ export type Apply = myra.Apply
 export interface ViewContext<TState, TProps> extends myra.ViewContext<TState, TProps> { }
 export interface Render<TState, TProps> extends myra.Render<TState, TProps> { }
 
-export interface EventListener<T extends Event, E extends Element> extends myra.EventListener<T, E> { }
+export type EventListener<T extends Event, E extends Element> = myra.EventListener<T, E>
+export interface EventListenerReturningVoid<TEvent extends Event, TElement extends Element> extends myra.EventListenerReturningVoid<TEvent, TElement> { }
+export interface EventListenerReturningState<TEvent extends Event, TElement extends Element> extends myra.EventListenerReturningState<TEvent, TElement> { }
 
 export interface VNodeBase extends myra.VNodeBase { }
 export interface TextVNode extends myra.TextVNode { }
